fix: Consolidate API endpoints to single file for Vercel Hobby plan
- Merge all feedback API endpoints into single index.js file
- Remove separate API files to stay within 12 function limit
- Consolidate Firebase initialization and CRUD operations
- Maintain all functionality: comments, suggestions, stats, contact
- Ready for deployment within Vercel Hobby plan limits
